refactoring menu
